Move src/hardware/parport to src/hardware/printer
The directory only ever held the virtual printer code; calling it
'parport' was a hangover from upstream DOSBox's parallel-port
subsystem (which has no analogue in dosbox-staging). Renaming makes
the directory's purpose match its contents and follows the audio/
input/ naming used elsewhere in src/hardware/.

Pure rename: same files, same content. CMake source paths and the
#include in dosbox.cpp are updated to match.
