+++ src/hardware/printer/printer.cpp
