+++ src/hardware/printer/printer_glue.cpp
