+++ src/hardware/printer/printer_glue.h
